@@ -1,48 +1,47 @@
 {
-    "name": "my-flix-angular-client",
-    "version": "0.0.0",
-    "scripts": {
-      "ng": "ng",
-      "start": "ng serve",
-      "build": "ng build",
-      "watch": "ng build --watch --configuration development",
-      "test": "ng test"
-    },
-    "private": true,
-    "dependencies": {
-      "@angular/animations": "^16.2.12",
-      "@angular/cdk": "~17.3.10", 
-      "@angular/common": "^16.2.12",
-      "@angular/compiler": "^16.2.12",
-      "@angular/core": "^16.2.12",
-      "@angular/forms": "^16.2.12",
-      "@angular/material": "~17.3.10", 
-      "@angular/platform-browser": "^16.2.12",
-      "@angular/platform-browser-dynamic": "^16.2.12",
-      "@angular/platform-server": "^16.2.12",
-      "@angular/router": "^16.2.12",
-      "@nguniversal/express-engine": "^16.2.0",
-      "express": "^4.18.2",
-      "rxjs": "~7.8.0",
-      "tslib": "^2.3.0",
-      "zone.js": "~0.13.0" 
-    },
-    "devDependencies": {
-      "@angular-devkit/build-angular": "~16.2.12",
-      "@angular/cli": "^16.2.12",
-      "@angular/compiler-cli": "^16.2.12",
-      "@types/angular": "^1.8.9",
-      "@types/express": "^4.17.17",
-      "@types/jasmine": "~5.1.0",
-      "@types/jest": "^29.5.13",
-      "@types/node": "^20.0.0",
-      "jasmine-core": "~5.2.0",
-      "karma": "~6.4.0",
-      "karma-chrome-launcher": "~3.2.0",
-      "karma-coverage": "~2.2.0",
-      "karma-jasmine": "~5.1.0",
-      "karma-jasmine-html-reporter": "~2.1.0",
-      "typescript": "^5.4.5"
-    }
+  "name": "my-flix-angular-client",
+  "version": "0.0.0",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "^18.2.9",
+    "@angular/cdk": "~17.3.10",
+    "@angular/common": "^18.2.9",
+    "@angular/compiler": "^18.2.9",
+    "@angular/core": "^18.2.9",
+    "@angular/forms": "^18.2.9",
+    "@angular/material": "~17.3.10",
+    "@angular/platform-browser": "^18.2.9",
+    "@angular/platform-browser-dynamic": "^18.2.9",
+    "@angular/platform-server": "^18.2.9",
+    "@angular/router": "^18.2.9",
+    "@angular/ssr": "^18.2.10",
+    "express": "^4.18.2",
+    "rxjs": "~7.8.0",
+    "tslib": "^2.3.0",
+    "zone.js": "^0.14.10"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~18.2.10",
+    "@angular/cli": "^18.2.10",
+    "@angular/compiler-cli": "^18.2.9",
+    "@types/angular": "^1.8.9",
+    "@types/express": "^4.17.17",
+    "@types/jasmine": "~5.1.0",
+    "@types/jest": "^29.5.13",
+    "@types/node": "^20.0.0",
+    "jasmine-core": "~5.2.0",
+    "karma": "~6.4.0",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.1.0",
+    "typescript": "^5.4.5"
   }
-  
+}
